refactoring towards new structure
