--- src/cards/mod.rs
@@ -1,7 +1,0 @@
-pub mod card;
-
-// Definitions:
-//
-// * A `Pile` is any collection of cards, in a specific order.
-// * A `Cardpack` is a `Pile` that represents a specific type of cards
-//   sorted with none removed.
